refactor: split add/from destructure
