+++ typescript/src/codeActions/custom/addDestructure/addSplittedDestructure.ts
@@ -1,31 +1,6 @@
-import { findChildContainingExactPosition, getChangesTracker, getPositionHighlights, isValidInitializerForDestructure, makeUniqueName } from '../../utils'
-import { CodeAction } from '../getCodeActions'
+import { findChildContainingExactPosition, getChangesTracker, getPositionHighlights, isValidInitializerForDestructure, makeUniqueName } from '../../../utils'
 
-const createDestructuredDeclaration = (initializer: ts.Expression, type: ts.TypeNode | undefined, declarationName: ts.BindingName) => {
-    if (!ts.isPropertyAccessExpression(initializer)) return
-
-    const propertyName = initializer.name.text
-    const { factory } = ts
-
-    const bindingElement = factory.createBindingElement(
-        undefined,
-        declarationName.getText() === propertyName ? undefined : propertyName,
-        declarationName.getText(),
-    )
-
-    return factory.createVariableDeclaration(
-        factory.createObjectBindingPattern([bindingElement]),
-        undefined,
-        type ? factory.createTypeLiteralNode([factory.createPropertySignature(undefined, factory.createIdentifier(propertyName), undefined, type)]) : undefined,
-        initializer.expression,
-    )
-}
-const addDestructureToVariableWithSplittedPropertyAccessors = (
-    node: ts.Node,
-    sourceFile: ts.SourceFile,
-    formatOptions: ts.FormatCodeSettings | undefined,
-    languageService: ts.LanguageService,
-) => {
+export default (node: ts.Node, sourceFile: ts.SourceFile, formatOptions: ts.FormatCodeSettings | undefined, languageService: ts.LanguageService) => {
     const isValidInitializer = ts.isVariableDeclaration(node.parent) && node.parent.initializer && isValidInitializerForDestructure(node.parent.initializer)
 
     // Make sure it only triggers on the destructuring object or parameter
@@ -108,47 +83,3 @@ const addDestructureToVariableWithSplittedPropertyAccessors = (
         ],
     }
 }
-export default {
-    id: 'addDestruct',
-    name: 'Add Destruct',
-    kind: 'refactor.rewrite.add-destruct',
-    tryToApply(sourceFile, position, _range, node, formatOptions, languageService) {
-        if (!node || !position) return
-        const initialDeclaration = ts.findAncestor(node, n => ts.isVariableDeclaration(n)) as ts.VariableDeclaration | undefined
-
-        if (initialDeclaration && !ts.isObjectBindingPattern(initialDeclaration.name)) {
-            const { initializer, type, name } = initialDeclaration
-
-            const result = addDestructureToVariableWithSplittedPropertyAccessors(node, sourceFile, formatOptions, languageService)
-
-            if (result) return result
-
-            if (!initializer || !isValidInitializerForDestructure(initializer)) return
-
-            const tracker = getChangesTracker(formatOptions ?? {})
-            const createdDeclaration = createDestructuredDeclaration(initializer, type, name)
-            if (createdDeclaration) {
-                tracker.replaceRange(
-                    sourceFile,
-                    {
-                        pos: initialDeclaration.pos + initialDeclaration.getLeadingTriviaWidth(),
-                        end: initialDeclaration.end,
-                    },
-                    createdDeclaration,
-                )
-
-                const changes = tracker.getChanges()
-                if (!changes) return undefined
-                return {
-                    edits: [
-                        {
-                            fileName: sourceFile.fileName,
-                            textChanges: changes[0]!.textChanges,
-                        },
-                    ],
-                }
-            }
-        }
-        return addDestructureToVariableWithSplittedPropertyAccessors(node, sourceFile, formatOptions, languageService)
-    },
-} satisfies CodeAction
