+++ typescript/src/codeActions/custom/addDestructure/createDestructuredDeclaration.ts
@@ -1,0 +1,19 @@
+export default (initializer: ts.Expression, type: ts.TypeNode | undefined, declarationName: ts.BindingName) => {
+    if (!ts.isPropertyAccessExpression(initializer)) return
+
+    const propertyName = initializer.name.text
+    const { factory } = ts
+
+    const bindingElement = factory.createBindingElement(
+        undefined,
+        declarationName.getText() === propertyName ? undefined : propertyName,
+        declarationName.getText(),
+    )
+
+    return factory.createVariableDeclaration(
+        factory.createObjectBindingPattern([bindingElement]),
+        undefined,
+        type ? factory.createTypeLiteralNode([factory.createPropertySignature(undefined, factory.createIdentifier(propertyName), undefined, type)]) : undefined,
+        initializer.expression,
+    )
+}
